+++ PezetTensors/Histogram.h
@@ -3,4 +3,4 @@
 
 
 EXPORTED_METHOD int* GetHistogram(int width, int height, double* inValues);
-EXPORTED_METHOD int* Get2DHistogram(int width, int height, double* inValues);
+EXPORTED_METHOD double* Get2DHistogram(int width, int height, double* inValues);
